Split tests into per-set files with WATERLILY_TEST selector
  Replace maintests.jl with one file per set (+ helper.jl);
  runtests.jl selects sets via WATERLILY_TEST and picks
  the suite by (KernelAbstractions -> main, SIMD -> alloc).
